Refactor: 설명용 예시 파일 정리
